backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-multipart
pypdf
python-docx
pytesseract
Pillow
pymupdf
fpdf2
# DB
sqlalchemy
alembic
oracledb
# Auth
passlib[bcrypt]
bcrypt<5
python-jose[cryptography]
email-validator